requirements.txt
langchain-core
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
openai==1.40.6
huggingface_hub==0.25.1
transformers==4.53.0
langchain==0.2.14
langchain-openai==0.1.22
httpx==0.28
pyjwt==2.9.0
slowapi==0.1.9
limits==3.11.0